requirements.txt
jsonschema-4.24.0
pytest==8.4.1
pytest-html-4.1.1
pytest-xdist-3.8.0
requests==2.32.4
